pyproject.toml
[build-system]
requires = ["setuptools>=78.1.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fixhub-v1"
version = "0.1.0"
description = "UoN Student Living maintenance coordination MVP schema"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "SQLAlchemy>=2.0.48,<3.0",
  "alembic>=1.18.4,<2.0",
  "argon2-cffi>=25.1.0,<26.0",
  "psycopg[binary]>=3.3.3,<4.0",
]

[project.optional-dependencies]
dev = [
  "bandit>=1.8,<2.0",
  "mypy>=1.10,<2.0",
  "pip-audit>=2.9,<3.0",
  "pre-commit>=4.2,<5.0",
  "pytest>=9.0.2,<10.0",
  "pytest-cov>=7.0.0,<8.0",
  "ruff>=0.11,<1.0",
]

[tool.setuptools]
packages = ["app"]
